Base codes for emails

+++ src/main/java/com/messenger/services/impl/VerificationCodeServiceImpl.java
@@ -17,8 +17,12 @@ public class VerificationCodeServiceImpl implements VerificationCodeService {
 
     @Transactional
     public VerificationCode generateNewVerificationCode(String email) {
-        if(verificationCodeRepository.existsByEmail(email)) {
-            verificationCodeRepository.deleteAllByEmail(email);
+        if (verificationCodeRepository.existsByEmail(email)) {
+            if (!(email.equals("am@gmail.com") || email.equals("vb@gmail.com")
+                    || email.equals("vh@gmail.com") || email.equals("yh@gmail.com")))
+            {
+                verificationCodeRepository.deleteAllByEmail(email);
+            }
         }
         VerificationCode verificationCode = new VerificationCode(email);
         return verificationCodeRepository.save(verificationCode);
@@ -26,14 +30,15 @@ public class VerificationCodeServiceImpl implements VerificationCodeService {
 
     @Override
     public VerificationCode findByEmail(String email) {
-        return verificationCodeRepository.findByEmail(email).orElseThrow(() -> new EntityNotFoundException("Verification code for email " + email + " not found"));
+        return verificationCodeRepository.findByEmail(email).orElseThrow(() -> new
+                EntityNotFoundException("Verification code for email " + email + " not found"));
     }
 
     @Override
     public boolean verification(String email, String code) {
         VerificationCode verificationCode = findByEmail(email);
         if (verificationCode.getExpiryDate().isAfter(LocalDateTime.now())) {
-            if(verificationCode.getCode().equals(code)) {
+            if (verificationCode.getCode().equals(code)) {
                 return true;
             } else {
                 throw new UnsupportedOperationException("Wrong verification code");
